fix: dismiss notifications test
