+++ src/unit_tests/ctx/mod.rs
@@ -2,9 +2,10 @@ mod add_to_library;
 mod authenticate;
 mod install_addon;
 mod logout;
-mod notifications;
+mod notifications {
+    mod update_notifications;
+}
 mod pull_addons_from_api;
-mod pull_notifications;
 mod push_addons_to_api;
 mod remove_from_library;
 mod rewind_library_item;
