--- src/unit_tests/ctx/notifications/mod.rs
@@ -1,1 +1,0 @@
-mod notifications;
